Refactoring reverseimagesearch
Removed unused functions and refactored some functions into searchimage.py
